bugfix field base data type & mul; impl transcript

+++ src/circuit.rs
@@ -76,7 +76,7 @@ impl Circuit {
                 let gate = GateMul {
                     i_ids: [mul_input[i * 4 + 1], mul_input[i * 4 + 2]],
                     o_id: mul_input[i * 4 + 3],
-                    coef: FPrimitive::from(mul_input[i * 4 + 4]),
+                    coef: FPrimitive::from(mul_input[i * 4 + 4] as u32),
                 };
                 layer.mul.push(gate);
             }
@@ -95,7 +95,7 @@ impl Circuit {
                 let gate = GateAdd {
                     i_ids: [add_input[i * 3 + 1]],
                     o_id: add_input[i * 3 + 2],
-                    coef: FPrimitive::from(add_input[i * 3 + 3]),
+                    coef: FPrimitive::from(add_input[i * 3 + 3] as u32),
                 };
                 layer.add.push(gate);
             }
